Refactor gen-assets and generate site.json
